day 8.1 elixir

+++ bootstrap.sh
@@ -1,11 +1,15 @@
 set -e
 
-source ./.aoc_session.sh
+if [[ -f "bootstrap.sh" ]]; then
+    echo "Please run from same directory"
+fi
 
 if [[ -z ${1} ]]; then
     echo 'Supply day'
     exit
 fi
+
+AOC_SESSION=$(cat .aoc_session)
 
 year="2021"
 day="$1"
@@ -20,10 +24,13 @@ if [[ ! -d "$daydir" ]]; then
 fi
 
 if [[ ! -f "$daydir/input.txt" ]]; then
-    curl https://adventofcode.com/$year/day/$day/input \
+    curl -s https://adventofcode.com/$year/day/$day/input \
         --cookie "session=$AOC_SESSION" \
         -o "$daydir/input.txt"
 fi
 
 touch "$daydir/sample.txt"
-cp template.py "$daydir/day$day.py"
+
+if [[ ! -f "$daydir/day$day.py" ]]; then
+    cp template.py "$daydir/day$day.py"
+fi
